+++ workspaces/scorecard/plugins/scorecard/src/utils/chartLabelUtils.ts
@@ -1,0 +1,37 @@
+/*
+ * Copyright Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export const getYOffsetForCenterLabel = (lineCount: number) => {
+  switch (lineCount) {
+    case 2:
+      return -17;
+    case 3:
+      return -24;
+    default:
+      return -8;
+  }
+};
+
+export const getHeightForCenterLabel = (lineCount: number) => {
+  switch (lineCount) {
+    case 2:
+      return 48;
+    case 3:
+      return 56;
+    default:
+      return 40;
+  }
+};
